# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Response data from the Samehadaku upstream API
 */
export interface AnimeResponse {
  [key: string]: unknown;
}

export interface UpstreamErrorBody {
  error: string;
  message: string;
}

/**
 * Failed to fetch from upstream API
 */
export type UpstreamErrorResponse = UpstreamErrorBody;

export type AnimeSearchParams = {
  /**
   * Search query
   */
  q: string;
};