# lib/api-zod/src/generated/types/animeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Response data from the Samehadaku upstream API
 */
export interface AnimeResponse {
  [key: string]: unknown;
}